[designators] moved parent designator descriptions to designator.py
