+++ src/pycram/designators/action_designator.py
@@ -16,102 +16,9 @@ from ..orm.base import Quaternion, Position, Base, RobotState, MetaData
 from ..robot_descriptions.robot_description_handler import InitializedRobotDescription as robot_description
 from ..task import with_tree
 from ..enums import Arms
+from ..designator import ActionDesignatorDescription
 from ..bullet_world import BulletWorld
 
-
-class ActionDesignatorDescription(DesignatorDescription):
-    """
-    Abstract class for action designator descriptions.
-    Descriptions hold possible parameter ranges for action designators.
-    """
-
-    @dataclasses.dataclass
-    class Action:
-        """
-        The performable designator with a single element for each list of possible parameter.
-        """
-        robot_position: Tuple[List[float], List[float]] = dataclasses.field(init=False)
-        """
-        The position of the robot at the start of the action.
-        """
-        robot_torso_height: float = dataclasses.field(init=False)
-        """
-        The torso height of the robot at the start of the action.
-        """
-
-        robot_type: str = dataclasses.field(init=False)
-        """
-        The type of the robot at the start of the action.
-        """
-
-        def __post_init__(self):
-            self.robot_position = BulletWorld.robot.get_position_and_orientation()
-            self.robot_torso_height = BulletWorld.robot.get_joint_state(robot_description.i.torso_joint)
-            self.robot_type = BulletWorld.robot.type
-
-        @with_tree
-        def perform(self) -> Any:
-            """
-            Executes the action with the single parameters from the description.
-            """
-            raise NotImplementedError()
-
-        def to_sql(self) -> ORMAction:
-            """
-            Create an ORM object that corresponds to this description.
-
-            :return: The created ORM object.
-            """
-            raise NotImplementedError(f"{type(self)} has no implementation of to_sql. Feel free to implement it.")
-
-        def insert(self, session: sqlalchemy.orm.session.Session, *args, **kwargs) -> ORMAction:
-            """
-            Add and commit this and all related objects to the session.
-            Auto-Incrementing primary keys and foreign keys have to be filled by this method.
-
-            :param session: Session with a database that is used to add and commit the objects
-            :param args: Possible extra arguments
-            :param kwargs: Possible extra keyword arguments
-            :return: The completely instanced ORM object
-            """
-
-            # get or create metadata
-            metadata = MetaData().insert(session)
-
-            # create position
-            position = Position(*self.robot_position[0])
-            position.metadata_id = metadata.id
-
-            # create orientation
-            orientation = Quaternion(*self.robot_position[1])
-            orientation.metadata_id = metadata.id
-
-            session.add_all([position, orientation])
-            session.commit()
-
-            # create robot-state object
-            robot_state = RobotState()
-            robot_state.position = position.id
-            robot_state.orientation = orientation.id
-            robot_state.torso_height = self.robot_torso_height
-            robot_state.type = self.robot_type
-            robot_state.metadata_id = metadata.id
-            session.add(robot_state)
-            session.commit()
-
-            # create action
-            action = self.to_sql()
-            action.metadata_id = metadata.id
-            action.robot_state = robot_state.id
-
-            return action
-
-    def __init__(self, resolver=None):
-        super(ActionDesignatorDescription, self).__init__(resolver)
-
-    def ground(self) -> Action:
-        """Fill all missing parameters and chose plan to execute. """
-        raise NotImplementedError(f"{type(self)}.ground() is not implemented.")
 
 
 class MoveTorsoAction(ActionDesignatorDescription):
